(RE-5460, RE-5449) Enable AIX builds
This is the changeset needed to build an AIO for AIX.

It adds AIX platform definitions for AIX 5.3, 6.1 and 7.1.

It also modifies components to build for AIX. In most cases, this is
just for items that compile. Often things like LDFLAGS and CFLAGS need
adjustments for AIX. We also carry a few patches for RUby on AIX.

For puppet and mco a service definition had to be added.

This build has been tested on AIX 6.1 and AIX 7.1. (AIX 5.3 pending)